api_database/requirements.txt
flask
flask_sqlalchemy
flask_swagger_ui
flask_marshmallow
flask-cors
marshmallow-sqlalchemy